Add font_size config
Also extract some code to `commands`.
