+++ config.json
@@ -1,4 +1,5 @@
 {
+  "font_size": 14,
   "commands": {
     "key": "",
     "name": "Root",
